refactor(engine): rename files for consistency
