+++ packages/engine/src/domain/action/action_type.ts
@@ -1,0 +1,6 @@
+export const ActionType = {
+  MOVE: "move",
+  SPLIT_MOVE: "split_move",
+} as const;
+
+export type ActionType = (typeof ActionType)[keyof typeof ActionType];
