+++ packages/engine/src/domain/action/interfaces.ts
@@ -1,11 +1,5 @@
+import type { ActionType } from "@/domain/action/action_type";
 import type { ICoordinate } from "@/math/coordinate";
-
-export const ActionType = {
-  MOVE: "move",
-  SPLIT_MOVE: "split_move",
-} as const;
-
-export type ActionType = (typeof ActionType)[keyof typeof ActionType];
 
 export interface IAction {
   readonly playerId: string;
