+++ packages/engine/src/domain/game/game.ts
@@ -1,11 +1,11 @@
-import type { IAction } from "@/domain/action/action";
+import type { IAction } from "@/domain/action/interfaces";
 import type { GameMode } from "@/domain/game/game-mode";
 import type { IGameResult } from "@/domain/game/game-result";
 import type { GameStatus } from "@/domain/game/game-status";
-import type { IGrid } from "@/domain/grid/grid";
-import type { IItem } from "@/domain/item/item";
-import type { IPlayer } from "@/domain/player/player";
-import type { Team } from "@/domain/team/team";
+import type { IGrid } from "@/domain/grid/interfaces";
+import type { IItem } from "@/domain/item/interfaces";
+import type { IPlayer } from "@/domain/player/interfaces";
+import type { Team } from "@/domain/team/interfaces";
 import type { ICoordinate } from "@/math/coordinate";
 
 /**
