+++ packages/engine/src/domain/grid/interfaces.ts
