+++ packages/engine/src/domain/item/interfaces.ts
@@ -1,10 +1,4 @@
-export const ItemType = {
-  BOMB: 0, // Demolition mode
-  RUGBY_BALL: 1, // Rugby mode
-  FLAG: 2, // CTF mode
-} as const;
-
-export type ItemType = (typeof ItemType)[keyof typeof ItemType];
+import type { ItemType } from "@/domain/item/item-type";
 
 /**
  * Represents a carryable object on the grid.
